requirements.txt
Django~=3.1.1
pip~=21.1.3
attrs~=21.2.0
pytz~=2020.1
idna~=2.10
multidict~=5.1.0
Babel~=2.9.1
sqlparse~=0.3.1
asgiref~=3.2.10
click~=7.1.2
aiohttp~=3.7.4.post0
certifi~=2021.5.30
aiogram~=2.13
aioredis~=1.3.1
yarl~=1.6.3
chardet~=3.0.4
setuptools~=49.2.1
python-dotenv~=0.17.1
psycopg2-binary~=2.9.1
gunicorn~=20.1.0
validators~=0.18.2
PyJWT~=2.1.0
requests~=2.25.0
hiredis~=2.0.0
urllib3~=1.26.6
six~=1.15.0
decorator~=5.1.0
phonenumbers~=8.12.35